# application/backend/certiblock/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/students": {
            "post": {
                "description": "Register a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Register a student",
                "parameters": [
                    {
                        "description": "Student data",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.StudentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.StudentOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/students/certificates": {
            "post": {
                "description": "Get all certificates of a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get all certificates of a student",
                "parameters": [
                    {
                        "description": "Student data",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.StudentAuth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.CertificateOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/students/certificates/{certUUID}": {
            "post": {
                "description": "Get one certificate of a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Get one certificate of a student",
                "parameters": [
                    {
                        "description": "Student data",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.StudentAuth"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Certificate UUID",
                        "name": "certUUID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.CertificateOutputFull"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/api/students/login": {
            "post": {
                "description": "Login a student",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "students"
                ],
                "summary": "Login a student",
                "parameters": [
                    {
                        "description": "Student data",
                        "name": "student",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.StudentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.StudentOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.CertificateOutput": {
            "type": "object",
            "properties": {
                "certHash": {
                    "type": "string"
                },
                "certUUID": {
                    "type": "string"
                },
                "dateOfIssuing": {
                    "type": "string"
                },
                "studentPublicKey": {
                    "type": "string"
                },
                "universityPublicKey": {
                    "type": "string"
                }
            }
        },
        "data.CertificateOutputFull": {
            "type": "object",
            "properties": {
                "base64File": {
                    "type": "string"
                },
                "certHash": {
                    "type": "string"
                },
                "certUUID": {
                    "type": "string"
                },
                "dateOfIssuing": {
                    "type": "string"
                },
                "studentPublicKey": {
                    "type": "string"
                },
                "universityPublicKey": {
                    "type": "string"
                }
            }
        },
...
        "data.StudentAuth": {
            "type": "object",
            "properties": {
                "publicKey": {
                    "type": "string"
                }
            }
        },
        "data.StudentInput": {
            "type": "object",
            "properties": {
                "NIN": {
                    "type": "string"
                },
                "countryID": {
                    "type": "integer"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                }
            }
        },
        "data.StudentOutput": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "privateKey": {
                    "type": "string"
                },
                "publicKey": {
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        }
    }
}